requirements.txt
Flask
python-docx
PyMuPDF
openai
python-dotenv
requests
groq